\docType{data}
\name{blackouts}
\alias{blackouts}
\title{Electrical blackout}
\format{A vector}
\source{
  M. E. J. Newman, "Power laws, Pareto distributions and
  Zipf's law." Contemporary Physics 46, 323 (2005). See
  http://tuvalu.santafe.edu/~aaronc/powerlaws/data.htm for
  further details.
}
\description{
  The numbers of customers affected in electrical blackouts
  in the United States between 1984 and 2002.
}
\keyword{datasets}

