% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_xmin.R
\name{get_KS_statistic}
\alias{get_KS_statistic}
\title{Deprecated function}
\usage{
get_KS_statistic(m, xmax = 1e+05, distance = "ks")
}
\arguments{
\item{m}{A reference class object that contains the data.}

\item{xmax}{default \code{1e5}. The maximum x value calculated when working out the CDF. See details for further 
information.}

\item{distance}{A string containing the distance measure (or measures) to calculate. 
Possible values are \code{ks} or \code{reweight}. See details for further information.}
}
\description{
This function is now deprecated and may be removed in future 
versions.
}
\seealso{
get_distance_statistic
}
