% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n.R
\name{get_n}
\alias{get_n}
\title{Sample size}
\usage{
get_n(m)
}
\arguments{
\item{m}{a distribution object.}
}
\description{
Returns the sample size of the data set contained within the 
distribution object.
}
\examples{
################################################
# Load data and create example object
################################################
data(moby_sample)
m = displ$new(moby_sample)
################################################
# get_n and length should return the same value
################################################
get_n(m)
length(moby_sample)
}

