% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{moby}
\alias{moby}
\alias{moby_sample}
\title{Moby Dick word count}
\format{
A vector
}
\source{
M. E. J. Newman, "Power laws, Pareto distributions and Zipf's law."
Contemporary Physics 46, 323 (2005).
See http://tuvalu.santafe.edu/~aaronc/powerlaws/data.htm for further details.
}
\description{
The frequency of occurrence of unique words in the novel Moby Dick
by Herman Melville.

The data set moby_sample is 2000 values
sampled from the moby data set.
}
