\docType{data}
\name{NativeAmerican}
\alias{NativeAmerican}
\alias{USAmerican}
\title{Casualities in the American Indian Wars (1776 and 1890)}
\format{Data frame}
\source{
  Friedman, Jeffrey A. "Using Power Laws to Estimate
  Conflict Size." (2013).
}
\description{
  These data files contain the observed casualties in the
  American Indian Wars. The data sets \code{NativeAmerican}
  and \code{USAmerican} contain the casualties on the
  Native American and US American sides respectively. Each
  data set is a data frame, with two columns: the number of
  casualities and the conflict date.
}
\keyword{datasets}

