% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.bs_xmin}
\alias{plot.bs_p_xmin}
\alias{plot.bs_xmin}
\alias{plot.compare_distributions}
\title{Plot methods for bootstrap objects}
\usage{
\method{plot}{bs_xmin}(x, trim = 0.1, ...)

\method{plot}{bs_p_xmin}(x, trim = 0.1, ...)

\method{plot}{compare_distributions}(x, ...)
}
\arguments{
\item{x}{an object of class \code{bs_xmin} or \code{bs_p_xmin}}

\item{trim}{When plotting the cummulative means and standard deviation, the first trim percentage of values are not displayed.
default \code{trim=0.1}}

\item{...}{graphics parameters to be passed to the plotting routines.}
}
\description{
A simple wrapper around the plot function to aid with visualising the bootstrap results.
The values plotted are returned as an invisible object.
}

