% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{population}
\alias{Population}
\alias{population}
\title{City boundaries and the universality of scaling laws}
\format{vector}
\source{
Arcaute, Elsa, et al. "City boundaries and the universality of scaling laws."
arXiv preprint arXiv:1301.1674 (2013).
}
\description{
This data set contains the population size of cities and towns in England. For further details
on the algorithm used to determine city boundries, see the referenced paper.
}

