\docType{package}
\name{poweRlaw-package}
\alias{powerlaw}
\alias{poweRlaw}
\alias{powerlaw-package}
\alias{poweRlaw-package}
\title{The poweRlaw package}
\description{
  This package implements both the discrete and continuous
  maximum likelihood estimators for fitting the power-law
  distribution to data. Additionally, a goodness-of-fit
  based approach is used to estimate the lower cutoff for
  the scaling region.
}
\details{
  The code developed in this package has been heavily
  influnenced from the python and R code found at:
  \url{http://tuvalu.santafe.edu/~aaronc/powerlaws/} . In
  particular, the R code of Laurent Dubroca.
}
\author{
  \email{colin.gillespie@newcastle.ac.uk}
}
\seealso{
  \url{https://github.com/csgillespie/poweRlaw}
}
\keyword{package}

