% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,distribution-method}
\alias{show,distribution-method}
\title{Generic show method for distribution objects}
\usage{
\S4method{show}{distribution}(object)
}
\arguments{
\item{object}{A distribution object.}
}
\description{
The distribution objects have an internal structure that is used for caching purposes.
Using the default \code{show} method gives the illusion of duplicate values.
This show method aims to avoid this confusion.
}

