% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{native_american}
\alias{NativeAmerican}
\alias{USAmerican}
\alias{native_american}
\alias{us_american}
\title{Casualities in the American Indian Wars (1776 and 1890)}
\format{Data frame}
\source{
Friedman, Jeffrey A. "Using Power Laws to Estimate Conflict Size." The Journal of conflict resolution (2014).
}
\description{
These data files contain the observed casualties in the American Indian Wars. The data sets
\code{native_american} and \code{us_american} contain the casualties on the Native American and US American
sides respectively. Each data set is a data frame, with two columns: the number of casualities and the
conflict date.
}

