\docType{package}
\name{poweRlaw-package}
\alias{powerlaw}
\alias{poweRlaw}
\alias{powerlaw-package}
\alias{poweRlaw-package}
\title{The poweRlaw package}
\description{
  This paper poweRlaw package aims to make fitting power
  laws and other heavy-tailed distributions
  straightforward. This package contains R functions for
  fitting, comparing and visualising heavy tailed
  distributions. Overall, it provides a principaled
  approach to fitting power laws to data.
}
\details{
  The code developed in this package has been heavily
  influnenced from the python and R code found at:
  \url{http://tuvalu.santafe.edu/~aaronc/powerlaws/} . In
  particular, the R code of Laurent Dubroca.
}
\author{
  \email{colin.gillespie@newcastle.ac.uk}
}
\seealso{
  \url{https://github.com/csgillespie/poweRlaw}
}
\keyword{package}

