% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{poweRlaw-package}
\alias{poweRlaw}
\alias{poweRlaw-package}
\alias{powerlaw}
\alias{powerlaw-package}
\title{The poweRlaw package}
\description{
The poweRlaw package aims to make fitting power laws and other heavy-tailed distributions
straightforward. This package contains R functions for fitting, comparing and visualising heavy
tailed distributions. Overall, it provides a principaled approach to fitting power laws to data.
}
\details{
The code developed in this package has
been heavily influnenced from the python and R code found at:
\url{http://tuvalu.santafe.edu/~aaronc/powerlaws/} .
In particular, the R code of Laurent Dubroca and Cosma Shalizi.
}
\author{
\email{colin.gillespie@newcastle.ac.uk}
}
\references{
A. Clauset, C.R. Shalizi, and M.E.J. Newman,
"Power-law distributions in empirical data" SIAM Review 51(4), 661-703 (2009).
}
\seealso{
\url{https://github.com/csgillespie/poweRlaw}
}
\keyword{package}

