% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw-package.R
\docType{package}
\name{poweRlaw-package}
\alias{poweRlaw-package}
\alias{poweRlaw}
\alias{powerlaw}
\alias{powerlaw-package}
\title{The poweRlaw package}
\description{
The poweRlaw package aims to make fitting power laws and other heavy-tailed distributions
straightforward. This package contains R functions for fitting, comparing and visualising heavy
tailed distributions. Overall, it provides a principled approach to fitting power laws to data.
}
\details{
The code developed in this package has
been heavily influenced  by the R code of Laurent Dubroca and Cosma Shalizi.
}
\references{
A. Clauset, C.R. Shalizi, and M.E.J. Newman,
"Power-law distributions in empirical data" SIAM Review 51(4), 661-703 (2009).
}
\seealso{
\url{https://github.com/csgillespie/poweRlaw}
}
\author{
\email{colin.gillespie@newcastle.ac.uk}
}
\keyword{package}
