% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{swiss_prot}
\alias{Swiss_prot}
\alias{swiss_prot}
\title{Word frequency in the Swiss-Prot database}
\format{data frame}
\source{
Bell, MJ, Gillespie, CS, Swan, D, Lord, P. 
An approach to describing and analysing bulk biological annotation 
quality: A case study using UniProtKB. 
Bioinformatics 2012, 28, i562-i568.
}
\description{
This dataset contains all the words extracted from the 
Swiss-Prot version 9 data (with the resulting frequency for each word).
Other datasets for other database versions can be obtained by contacting 
Michael Bell 
(http://homepages.cs.ncl.ac.uk/m.j.bell1/annotationQualityPaper.php) 

Full details in http://arxiv.org/abs/arXiv:1208.2175v1
}

